Fixed halting on G1
Need to fix delay at end of G1 for proper ending/final position calculation/minimum speed change

+++ src/Kinematics/IPathPlanner.h
@@ -5,7 +5,7 @@
 class IPathPlanner {
 public:
     virtual void AddAxis(Axis* axis) = 0;
-    virtual Axis* GetAxis(int axisIndex) = 0;
+    virtual Axis* GetAxis(uint8_t axisIndex) = 0;
     virtual uint8_t GetAxisCount() = 0;
     virtual void CalculateLimits(float feedrate) = 0;
     virtual bool Update() = 0;
